picasso extension function now is located in a separate file
